Lookup whether there are child links in Links queries
This consolidates the 2 Queries::Links* files into a single
Queries::Link class that has two static methods. This is because there
is now a good amount of shared logic between them.

This adds the ability for the query that is checking for links to check
whether any of the links it is retrieving has links, this can be used to
save a query that pulls out 0 data, and can save us hundreds of queries
in the process of building a link graph.

The downside though is that due to the complexities of link expansion and
dependency resolution it has hard terminology to understand. The phrasing
of from and to links is used rather than direct and reverse, as the
latter means different things in link expansion and dependency
resolution.