Make children a concern of a LinkGraph::Node
We can now look up whether a link has child links at time we are pulling
it from the db. The changes here allow this information to flow into a
LinkGraph::Node and can be used to determine what is needed to do at the
next stage of looking up the children.

It has two fields `has_own_links` and `is_linked_to`.

- `has_own_links` is used to indicate that the content_id of this node has
a LinkSet which contains links to other content_ids of a subset of
link_types
- `is_linked_to` indicates that there are Links where the
target_content_id is the nodes content_id of a subset of link_types.

The subset of link_types is determined by the ExpansionRules and differ
depending on the circumstance of the links. What are allowed link_types
is not a concern of this class.